+++ fetch_data/sparql_data/connectivity_math/__init__.py
@@ -1,0 +1,8 @@
+"""Calculations for Connectivity Observations"""
+
+from fetch_data.sparql_data.connectivity_math.connectivity_distance_dictionary import (
+    compile_distance_dict,
+)
+from fetch_data.sparql_data.connectivity_math.connectivity_link_dictionary import (
+    compile_link_dict,
+)
